Crate http_server

Simple example starting an Oak Application serving a static page over HTTP. In addition, there is a redirect server that redirects the requests sent to it to the static server. This is done using an HTTP client pseudo-Node. So, in addition to the Main node and the Log node, this example has three Oak nodes, and three pseudo-Nodes.

Modules

proto

Structs

RedirectHandler

A handler corresponding to an HTTP server that redirects every request to https://localhost:8080 and responds with OK (200) if the redirect is successful.

RedirectHelper

A helper node that sends request to an HTTP client pseudo-Node, collects the response and logs the status in the response. To be able to interact with the HTTP client pseudo-Node, this node expects to receive an oak::http::HttpInvocationSource in the invocation it receives for each command.

StaticHttpHandler

A simple HTTP handler that responds with OK (200) to every request sent to /, and with NOT_FOUND (400) to any other request. This struct is used in the abitest. So its functionality should be modified with care!

Functions

http_handler

oak_main

redirect_handler

redirect_helper